Make getNetworkInterfaces() and getBestPublicIp() available via tools.h Remove HTTP URL helper line - should be done in kiwix-serve Add getters at server level - getAddress and getPort
